Moved packages 'alarm' and 'location' in the common super package 'managing', along with the classes "manager", "MessageBuilder" and "Service Manager"
Did some refactoring due to the Presence of MessageBuilder
