+++ app/src/main/java/ingsw/group1/findmyphone/managing/location/Command.java
@@ -1,4 +1,4 @@
-package ingsw.group1.findmyphone.location;
+package ingsw.group1.findmyphone.managing.location;
 
 /**
  * Interface built to make method getLastLocation modular,
